Adds ellipse back in after content limit
includes a filter on `get_the_content_limit` if the content limit is
used in the widget. The filter is added before the output and removed
after the output to ensure the output is not affected outside the
widget.

Uses str_replace to insert an ellipse before the closing </p> of the
content limit output.